+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
+## [unreleased]
+
+- Color subtest variations
+- Keep the ':' as is, but color it
+
 ## [0.1.1] - 2022-03-21
 
 - Remove colon after --- FAIL: